win32: fix mingw wrapper regressions

+++ contrib/win32/win32compat/signal_sigalrm.c
@@ -46,7 +46,7 @@ sigalrm_APC(_In_opt_ LPVOID lpArgToCompletionRoutine,
 }
 
 unsigned int
-w32_alarm(long long sec)
+w32_alarm(unsigned int sec)
 {
 	LARGE_INTEGER due;
 	ULONGLONG sec_passed;
